requirements.txt
Django==2.2
django-cors-headers==2.5.3
django-richtextfield==1.4.0
djangorestframework==3.9.2
Pillow==6.0.0
pytz==2019.1
sqlparse==0.3.0
